pyproject.toml
[tool.poetry]
name = "FastAPI Clean Architecture"
version = "0.1.0"
description = ""
authors = ["bodaue"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
pydantic-settings = "^2.6.1"
dishka = "^1.4.1"
uvicorn = "^0.34.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.37" }
asyncpg = "^0.30.0"
bcrypt = "^4.2.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
alembic = "^1.14.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
ruff = "^0.8.0"
pre-commit = "^4.0.1"

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
follow_imports = "normal"
show_error_context = false
pretty = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = [
    "ALL"
]
ignore = [
    "RUF001", "RUF002", "RUF003", "RUF012",
    "COM812", "ISC001",
    "D",
    "RET502", "RET503",
    "ANN401",
    "I001", "INP001",
    "FBT001", "FBT002",
    "TRY003", "EM101"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
